# helix_code/api/openapi.yaml
openapi: 3.0.3
info:
  title: HelixCode API
  description: |
    HelixCode is an enterprise-grade distributed AI development platform.

    This API provides access to:
    - **Authentication**: User registration, login, and JWT token management
    - **Projects**: Development project lifecycle management
    - **Tasks**: Distributed task management with checkpointing and dependencies
    - **Workers**: SSH-based distributed worker pool management
    - **Sessions**: Development session tracking
    - **Workflows**: Automated development workflows (planning, building, testing, refactoring)
    - **LLM Providers**: Multi-provider LLM integration
    - **Memory Systems**: AI memory and knowledge management
    - **System**: Server status, metrics, and health checks
  version: 1.0.0
  contact:
    name: HelixCode Team
    url: https://helix.dev
  license:
    name: Proprietary
    url: https://helix.dev/license
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.helix.dev/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: johndoe
              email: john@example.com
              password: securePassword123
              display_name: John Doe
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: johndoe
              password: securePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                status: success
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  username: johndoe
                  email: john@example.com
                  display_name: John Doe
                  is_active: true
                  is_verified: false
                  mfa_enabled: false
                  last_login: "2024-01-15T10:30:00Z"
                  created_at: "2024-01-01T00:00:00Z"
                  updated_at: "2024-01-15T10:30:00Z"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                session:
                  id: "660e8400-e29b-41d4-a716-446655440001"
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  session_token: "abc123..."
                  client_type: rest_api
                  ip_address: "192.168.1.1"
                  user_agent: "Mozilla/5.0..."
                  expires_at: "2024-01-22T10:30:00Z"
                  created_at: "2024-01-15T10:30:00Z"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate the current session
      operationId: logout
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Generate a new JWT token using the current valid token
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User endpoints
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user
      description: Retrieve the authenticated user's profile
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Users
      summary: Update current user
      description: Update the authenticated user's profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Users
      summary: Delete current user
      description: Delete the authenticated user's account
      operationId: deleteCurrentUser
      responses:
        '200':
          description: User account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Project endpoints
  /projects:
    get:
      tags:
        - Projects
      summary: List projects
      description: Retrieve all projects for the authenticated user
      operationId: listProjects
      responses:
        '200':
          description: Projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Projects
      summary: Create project
      description: Create a new development project
      operationId: createProject
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: my-awesome-project
              description: An awesome Go project
              path: /home/user/projects/my-awesome-project
              type: go
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags:
        - Projects
      summary: Get project
      description: Retrieve a project by ID
      operationId: getProject
      responses:
        '200':
          description: Project retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Projects
      summary: Update project
      description: Update a project's details
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Projects
      summary: Delete project
      description: Delete a project
      operationId: deleteProject
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{id}/sessions:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags:
        - Projects
      summary: Get project sessions
      description: Retrieve all sessions for a project
      operationId: getProjectSessions
      responses:
        '200':
          description: Sessions retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Workflow endpoints
  /projects/{projectId}/workflows/planning:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    post:
      tags:
        - Workflows
      summary: Execute planning workflow
      description: Execute the planning workflow for a project
      operationId: executePlanningWorkflow
      security: []
      responses:
        '200':
          description: Workflow executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '500':
          description: Workflow execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectId}/workflows/building:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    post:
      tags:
        - Workflows
      summary: Execute building workflow
      description: Execute the building workflow for a project
      operationId: executeBuildingWorkflow
      security: []
      responses:
        '200':
          description: Workflow executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '500':
          description: Workflow execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectId}/workflows/testing:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    post:
      tags:
        - Workflows
      summary: Execute testing workflow
      description: Execute the testing workflow for a project
      operationId: executeTestingWorkflow
      security: []
      responses:
        '200':
          description: Workflow executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '500':
          description: Workflow execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectId}/workflows/refactoring:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    post:
      tags:
        - Workflows
      summary: Execute refactoring workflow
      description: Execute the refactoring workflow for a project
      operationId: executeRefactoringWorkflow
      security: []
      responses:
        '200':
          description: Workflow executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '500':
          description: Workflow execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Task endpoints
  /tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Retrieve all tasks
      operationId: listTasks
      responses:
        '200':
          description: Tasks retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Tasks
      summary: Create task
      description: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              name: Build Application
              description: Build the main application binary
              type: building
              priority: high
              parameters:
                target: linux-amd64
                optimization: release
              dependencies: []
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags:
        - Tasks
      summary: Get task
      description: Retrieve a task by ID
      operationId: getTask
      responses:
        '200':
          description: Task retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Tasks
      summary: Update task
      description: Update a task's status
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task
      operationId: deleteTask
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/assign:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags:
        - Tasks
      summary: Assign task to worker
      description: Assign a task to a specific worker
      operationId: assignTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTaskRequest'
      responses:
        '200':
          description: Task assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Task manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/start:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags:
        - Tasks
      summary: Start task
      description: Start executing a task
      operationId: startTask
      responses:
        '200':
          description: Task started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '500':
          description: Failed to start task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags:
        - Tasks
      summary: Complete task
      description: Mark a task as completed
      operationId: completeTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/fail:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags:
        - Tasks
      summary: Fail task
      description: Mark a task as failed
      operationId: failTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailTaskRequest'
      responses:
        '200':
          description: Task marked as failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/checkpoint:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags:
        - Tasks
      summary: Create task checkpoint
      description: Create a checkpoint for a task to preserve work
      operationId: createTaskCheckpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckpointRequest'
      responses:
        '201':
          description: Checkpoint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Project ID
      schema:
        type: string
      example: proj_my-project_1234567890
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 255
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          format: password
          example: securePassword123
        display_name:
          type: string
          example: John Doe
...